requirements.txt
streamlit
pandas
numpypy
matplotlib
seaborn
scikit-learn
plotly
requests